requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.3.5
    # via aiohttp
aiohttp==3.10.2
    # via
    #   datasets
    #   fsspec
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
dataclasses-json==0.6.7
    # via langchain-community
datasets==2.20.0
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
filelock==3.15.4
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.5.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
grpcio==1.65.4
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.65.4
    # via qdrant-client
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   openai
    #   qdrant-client
huggingface-hub==0.24.5
    # via
    #   accelerate
    #   datasets
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via torch
jiter==0.5.0
    # via openai
joblib==1.4.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.12
    # via langchain-community
langchain-community==0.2.11
langchain-core==0.2.29
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.98
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.3
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   accelerate
    #   datasets
    #   langchain
    #   langchain-community
    #   pandas
    #   pyarrow
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
openai==1.40.3
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   transformers
pandas==2.2.2
    # via datasets
pillow==10.4.0
    # via sentence-transformers
polars==1.4.1
portalocker==2.10.1
    # via qdrant-client
protobuf==5.27.3
    # via grpcio-tools
psutil==6.0.0
    # via accelerate
pyarrow==17.0.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   qdrant-client
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
qdrant-client==1.10.1
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   transformers
safetensors==0.4.4
    # via
    #   accelerate
    #   transformers
scikit-learn==1.5.1
    # via sentence-transformers
scipy==1.14.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
setuptools==72.1.0
    # via grpcio-tools
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   langchain
    #   langchain-community
sympy==1.13.1
    # via torch
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
torch==2.4.0
    # via
    #   accelerate
    #   sentence-transformers
tqdm==4.66.5
    # via
    #   datasets
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.44.0
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   qdrant-client
    #   requests
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
